requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-allauth==0.52.0
django-browser-reload==1.6.0
django-click==2.3.0
django-environ==0.9.0
django-tailwind==3.5.0
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.14
